+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `fix-dns-nic-allocation.sh` Ignition script to attach DNS servers to correct network interfaces.
+
 ## [3.1.1] - 2025-12-17
 
 ### Changed